requirements.txt
altgraph==0.17
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
cryptography==3.4.6
future==0.18.2
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
lxml==4.6.2
oauthlib==3.1.0
pefile==2019.4.18
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyOpenSSL==20.0.1
PySocks==1.7.1
python-docx==0.8.10
pytz==2020.4
pywin32-ctypes==0.2.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.11
webdriver-manager==3.2.2
win-inet-pton==1.1.0
wincertstore==0.2
